refactoed backend routing
